\name{MyArmadillo1-package}
\alias{MyArmadillo1-package}
\alias{MyArmadillo1}
\docType{package}
\title{
\packageTitle{MyArmadillo1}
}
\description{
\packageDescription{MyArmadillo1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MyArmadillo1}
\packageIndices{MyArmadillo1}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MyArmadillo1}

Maintainer: \packageMaintainer{MyArmadillo1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
